+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## [Unreleased]
+* Add a `@coq_struct "param"` attribute to specify the decreasing parameter name of fixpoints.
 * Add a tactic `rewrite_cast_exists_eval_eq` to simplify the use of the `cast_exists` axiom in proofs.
 * Name the arguments of the signatures.
 * Rename `obj_magic` as `cast`.